# docs/swagger.yaml
definitions:
  db.CardModel:
    properties:
      authToken:
        type: string
      cardType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      expiryMonth:
        type: string
      expiryYear:
        type: string
      id:
        type: string
      isApproved:
        type: boolean
      isDefault:
        type: boolean
      lastDigits:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - authToken
    - cardType
    - createdAt
    - deletedAt
    - expiryMonth
    - expiryYear
    - id
    - isApproved
    - isDefault
    - lastDigits
    - updatedAt
    - user
    - userId
    type: object
  db.Currency:
    enum:
    - NGN
    type: string
    x-enum-varnames:
    - CurrencyNgn
  db.DeploymentLogModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      deployment:
        $ref: '#/definitions/db.DeploymentModel'
      deploymentId:
        type: string
      id:
        type: string
      jobId:
        type: string
      message:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - deployment
    - deploymentId
    - id
    - jobId
    - message
    - updatedAt
    type: object
  db.DeploymentModel:
    properties:
      DeploymentLog:
        items:
          $ref: '#/definitions/db.DeploymentLogModel'
        type: array
      actor:
        type: string
      commitHash:
        type: string
      commitMessage:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoConnection:
        $ref: '#/definitions/db.RepoConnectionModel'
      repoConnectionId:
        type: string
      status:
        $ref: '#/definitions/db.DeploymentStatus'
      updatedAt:
        type: string
    required:
    - DeploymentLog
    - actor
    - commitHash
    - commitMessage
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoConnection
    - repoConnectionId
    - status
    - updatedAt
    type: object
  db.DeploymentStatus:
    enum:
    - IDLE
    - IN_PROGRESS
    - SUCCESSFUL
    - FAILED
    type: string
    x-enum-varnames:
    - DeploymentStatusIdle
    - DeploymentStatusInProgress
    - DeploymentStatusSuccessful
    - DeploymentStatusFailed
  db.GithubAccountConnectionModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      githubEmail:
        type: string
      githubId:
        type: string
      githubInstallationId:
        type: integer
      githubUsername:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - deletedAt
    - githubEmail
    - githubId
    - githubInstallationId
    - githubUsername
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceModel:
    properties:
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      from:
        type: string
      id:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      reference:
        type: string
      status:
        $ref: '#/definitions/db.InvoiceStatus'
      to:
        type: string
      totalPrice:
        type: number
      unitPrice:
        type: number
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - currency
    - deletedAt
    - description
    - from
    - id
    - productId
    - quantity
    - reference
    - status
    - to
    - totalPrice
    - unitPrice
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceStatus:
    enum:
    - PAID
    - UNPAID
    type: string
    x-enum-varnames:
    - InvoiceStatusPaid
    - InvoiceStatusUnpaid
  db.MachineModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      Network:
        items:
          $ref: '#/definitions/db.NetworkModel'
        type: array
      RepoConnection:
        items:
          $ref: '#/definitions/db.RepoConnectionModel'
        type: array
      containerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machineImage:
        type: string
      machineName:
        type: string
      machineStatus:
        $ref: '#/definitions/db.MachineStatus'
      owner:
        $ref: '#/definitions/db.UserModel'
      ownerId:
        type: string
      plan:
        $ref: '#/definitions/db.MachinePlanModel'
      planId:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - Network
    - RepoConnection
    - containerId
    - createdAt
    - deletedAt
    - id
    - machineImage
    - machineName
    - machineStatus
    - owner
    - ownerId
    - plan
    - planId
    - updatedAt
    type: object
  db.MachinePlanModel:
    properties:
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      cpu:
        type: string
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      hourlyRate:
        type: number
      id:
        type: string
      memory:
        type: string
      monthlyRate:
        type: number
      name:
        type: string
      updatedAt:
        type: string
    required:
    - Machine
    - cpu
    - createdAt
    - currency
    - deletedAt
    - hourlyRate
    - id
    - memory
    - monthlyRate
    - name
    - updatedAt
    type: object
  db.MachineStatus:
    enum:
    - CREATING
    - RUNNING
    - REBOOTING
    - SHUTTING_DOWN
    - SHUTDOWN
    type: string
    x-enum-varnames:
    - MachineStatusCreating
    - MachineStatusRunning
    - MachineStatusRebooting
    - MachineStatusShuttingDown
    - MachineStatusShutdown
  db.NetworkModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      destinationPort:
        type: integer
      hostName:
        type: string
      id:
        type: string
      ingressId:
        type: string
      listeningPort:
        type: integer
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      protocol:
        type: string
      serviceId:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - destinationPort
    - hostName
    - id
    - ingressId
    - listeningPort
    - machine
    - machineId
    - protocol
    - serviceId
    - updatedAt
    type: object
  db.RepoConnectionModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoId
    - repoName
    - updatedAt
    type: object
  db.SocialConnectionModel:
    properties:
      connectionId:
        type: string
      connectionType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - connectionId
    - connectionType
    - createdAt
    - deletedAt
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionModel:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      reference:
        type: string
      status:
        $ref: '#/definitions/db.TransactionStatus'
      type:
        $ref: '#/definitions/db.TransactionType'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - amount
    - createdAt
    - currency
    - deletedAt
    - description
    - id
    - reference
    - status
    - type
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionStatus:
    enum:
    - PENDING
    - SUCCESS
    - FAILED
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusSuccess
    - TransactionStatusFailed
  db.TransactionType:
    enum:
    - DEBIT
    - CREDIT
    type: string
    x-enum-varnames:
    - TransactionTypeDebit
    - TransactionTypeCredit
  db.UserModel:
    properties:
      Card:
        items:
          $ref: '#/definitions/db.CardModel'
        type: array
      GithubAccountConnection:
        items:
          $ref: '#/definitions/db.GithubAccountConnectionModel'
        type: array
      Invoice:
        items:
          $ref: '#/definitions/db.InvoiceModel'
        type: array
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      SocialConnection:
        items:
          $ref: '#/definitions/db.SocialConnectionModel'
        type: array
      Transaction:
        items:
          $ref: '#/definitions/db.TransactionModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    required:
    - Card
    - GithubAccountConnection
    - Invoice
    - Machine
    - SocialConnection
    - Transaction
    - createdAt
    - deletedAt
    - email
    - firstName
    - id
    - lastName
    - password
    - roles
    - updatedAt
    type: object
...
  lib.AuthTokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    required:
    - accessToken
    - refreshToken
    type: object
  services.DeployRepoArgs:
    properties:
      connectionId:
        type: string
      ref:
        type: string
    required:
    - connectionId
    - ref
    type: object
  types.Auth0UserArgs:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  types.LoginUserArgs:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.LoginUserResult:
    properties:
      tokens:
        $ref: '#/definitions/lib.AuthTokens'
    required:
    - tokens
    type: object
...
  types.RefreshAccessTokenArgs:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  types.RegisterUserArgs:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.RegisterUserResult:
    properties:
      Card:
        items:
          $ref: '#/definitions/db.CardModel'
        type: array
      GithubAccountConnection:
        items:
          $ref: '#/definitions/db.GithubAccountConnectionModel'
        type: array
      Invoice:
        items:
          $ref: '#/definitions/db.InvoiceModel'
        type: array
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      SocialConnection:
        items:
          $ref: '#/definitions/db.SocialConnectionModel'
        type: array
      Transaction:
        items:
          $ref: '#/definitions/db.TransactionModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      tokens:
        $ref: '#/definitions/lib.AuthTokens'
      updatedAt:
        type: string
    required:
    - Card
    - GithubAccountConnection
    - Invoice
    - Machine
    - SocialConnection
    - Transaction
    - createdAt
    - deletedAt
    - email
    - firstName
    - id
    - lastName
    - password
    - roles
    - tokens
    - updatedAt
    type: object
...
  types.UpdateUserArgs:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    required:
    - email
    type: object
info:
  contact:
    email: formatio@overal-x.org
    name: Formatio Team
  title: Formatio API
  version: "1.0"
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      operationId: loginUser
      parameters:
      - description: Login User
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.LoginUserArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginUserResult'
      tags:
      - auth
  /auth/refresh-access-token/:
    post:
      consumes:
      - application/json
      operationId: refreshAccessToken
      parameters:
      - description: Refresh Access Token
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.RefreshAccessTokenArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginUserResult'
      tags:
      - auth
  /auth/register/:
    post:
      consumes:
      - application/json
      operationId: registerUser
      parameters:
      - description: Register User
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.RegisterUserArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.RegisterUserResult'
      tags:
      - auth
  /auth/social-connection/:
    post:
      consumes:
      - application/json
      operationId: authSocialConnection
      parameters:
      - description: Auth0 User
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.Auth0UserArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginUserResult'
      tags:
      - auth
...
  /deployments/deploy:
    post:
      consumes:
      - application/json
      operationId: deployRepo
      parameters:
      - description: Deploy Repo
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/services.DeployRepoArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - deployments
...
  /user/profile/:
    get:
      consumes:
      - application/json
      operationId: getProfileUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.UserModel'
      tags:
      - user
    patch:
      consumes:
      - application/json
      operationId: updateProfileUser
      parameters:
      - description: Update User
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/db.UserModel'
      tags:
      - user
swagger: "2.0"